Implement prefetching and priority downloading of chunks
When opening a file, the file gets divided into chunks of
a configurable size. Currently, the default is 2 MiB.
Those chunks will then be downloaded in the background and
cached in binary object files on disk.

When a task needs a chunk that's still queued, this chunk gets
put into a priority queue and will be downloaded first before
the downloader continues to download the other chunks that
no task currently waits for.