Fix a bug that caused cached content to corrupt
For files spanning multiple chunks, the generated chunks, the chunks
would overlap and cause the file to not be exactly the same as on the
remote side. Of course, this leads to checksums not matching and for
example causes artifacts in video files.

This patch fixes chunk generation to no longer overlap, which in turn
now yields matching checksums for all files, regardless of how many
chunks they span.